+++ hydra-node/test/Hydra/Network/AuthenticateSpec.hs
@@ -28,9 +28,11 @@ spec = parallel $ do
         NetworkCallback
           { deliver = \msg ->
               atomically $ modifyTVar' receivedMessages (msg :)
+          , onConnectivity = const $ pure ()
           }
 
   msg <- runIO $ generate @(Message SimpleTx) arbitrary
+
   it "pass the authenticated messages around" $ do
     let receivedMsgs = runSimOrThrow $ do
           receivedMessages <- newTVarIO []
@@ -64,8 +66,9 @@ spec = parallel $ do
             aliceSk
             [bob]
             ( \NetworkCallback{deliver} _ -> do
+                deliver (Signed msg (sign aliceSk msg) alice)
                 deliver (Signed msg (sign bobSk msg) bob)
-                deliver (Signed unexpectedMessage (sign aliceSk unexpectedMessage) alice)
+                deliver (Signed unexpectedMessage (sign carolSk unexpectedMessage) carol)
             )
             (captureIncoming receivedMessages)
             $ \_ ->
@@ -73,7 +76,7 @@ spec = parallel $ do
 
           readTVarIO receivedMessages
 
-    receivedMsgs `shouldBe` [Authenticated msg bob]
+    receivedMsgs `shouldBe` [Authenticated msg bob, Authenticated msg alice]
 
   it "drop message coming from party with wrong signature" $ do
     let receivedMsgs = runSimOrThrow $ do